refactor: scheduleListItem 네이밍에 명명 분리 기준을 추가 적용
