+++ frontend/src/features/shared-schedule/ui/ExpiredSchedules/expiredScheduleList.css.ts
