+++ frontend/src/features/shared-schedule/ui/ExpiredSchedules/expiredScheduleListItem.css.ts
