+++ frontend/src/features/shared-schedule/ui/UnConfirmedSchedules/unconfirmedScheduleList.css.ts
